Add support for java.net proxy settings.
Various tools work on the basis of the http(s).proxyHost and .proxyPort system
properties used by the java.net classes. Developers are also used to these
properties. This change picks up those settings and configures the appropriate
Proxy instance for all requests unless a different proxy is set either on the
request or on the SOAPClient or RESTClient instance.