Make all exceptions runtime.
It's easy to get complacent in Groovy with regards to checked and unchecked
exceptions, but there are certain situations where the difference rears its
ugly head. In my case, Groovy was wrapping the HTTPClientException with an
UndeclaredThrowableException from a Spock test because the former is a checked
exception but `RESTClient.get()` doesn't declare that it throws it.

One solution is to make the affected methods declare that they throw the
exception, but I just prefer to make the exceptions runtime. Thus they don't
need to be declared.