Defer use of GitHub project properties.
The Gradle build cannot be imported into an IDE unless the `githubUsername` and
`githubPassword` project properties are defined. This kind of behaviour puts off
potential contributors, particularly as they may not realise why there is a
problem.

This fixes the issue by deferring the use of the GitHub properties until it's
known that they are actually required, i.e. if the integration tests are to be
run. In addition, if the properties are needed but not provided, the build now
generates an appropriate error message.